requirements.txt
docutils==0.16
lockfile==0.12.2
python-daemon==2.2.4
